feat: 공간 상세조회 완료

+++ src/main/java/one/dfy/bily/api/inquiry/dto/InquiryCreateRequest.java
@@ -18,7 +18,7 @@ public record InquiryCreateRequest(
         String content,
         LocalDateTime createdAt,
         InquiryStatus status,
-        Integer spaceId,
+        Long spaceId,
         String hostCompany,
         String spaceIdName
 ) {}
